refactor: calendar view와 viewmodel로 분리, view는 순수한 ui만 표현
